[chore] refactored to MapViewScreen
