+++ components/Map/index.tsx
@@ -5,18 +5,6 @@ import AddMarkers from '../../api/maps/AddMarkers';
 import { Project } from '../../types/schema';
 import './styles.css';
 import { CSSProperties } from 'react';
-import { FaMapMarkerAlt } from 'react-icons/fa';
-import { FaBolt } from 'react-icons/fa6';
-import { IoIosCheckmarkCircleOutline } from 'react-icons/io';
-import { MdLightbulbOutline } from 'react-icons/md';
-import { FilterBar } from '../FilterBar';
-import { SearchBar } from '../SearchBar';
-
-interface Filter {
-  id: string;
-  label: string;
-  icon: React.ReactNode;
-}
 
 const containerStyle: CSSProperties = {
   width: '100%',
@@ -33,22 +21,9 @@ const center = {
 
 const mapId = '54eb1c7baba5a715'; // needed for AdvancedMarker
 
-const filters = [
-  { id: 'status', label: 'STATUS', icon: <IoIosCheckmarkCircleOutline /> },
-  { id: 'technology', label: 'TECHNOLOGY', icon: <FaBolt /> },
-  { id: 'projectSize', label: 'PROJECT SIZE', icon: <MdLightbulbOutline /> },
-  { id: 'location', label: 'LOCATION', icon: <FaMapMarkerAlt /> },
-];
-
-const handleFilterChange = (filter: Filter) => {
-  console.log(filter);
-};
-
 export default function Map(props: { projects: Project[] | null }) {
   return (
     <APIProvider apiKey={process.env.NEXT_PUBLIC_GOOGLE_MAPS_API_KEY as string}>
-      <SearchBar />
-      <FilterBar filters={filters} onFilterChange={handleFilterChange} />
       <GoogleMap
         style={containerStyle}
         defaultCenter={center}
